// src/app/models/vehicle.ts
export interface Vehicle {
  id: number;
  dateAdded: string;
  make: number;
  mileage: number;
  color: number;
  condition: number;
  modal: number;
  year: number;
  fuel: number;
  transmission: number;
  price: string;
  description: string;
  agentName: string;
  agentPhoneNumber: string;
  agentLocation: string;
  images: any[];
  userId: number;
}